Build a Gate Role and Policy

+++ app/Http/Controllers/Dashboard/CategoriesController.php
@@ -7,6 +7,7 @@ use App\Models\Category;
 use App\Rules\FilterRule;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
@@ -24,6 +25,12 @@ class CategoriesController extends Controller
         'image' => 'The art file should be image type'
     ];
 
+    # if you need exicute all policies
+    public function __construct()
+    {
+        $this->authorizeResource(Category::class);
+    }
+
     // in here build the Action Methods for route
     public function index()
     {
@@ -35,11 +42,21 @@ class CategoriesController extends Controller
         // dd($entries[0]);
         // dd($entries);
 
-        // $categories = Category::get();
 
         //todo using (inner-join) for relations to get the data from the two table
         // $categories = Category::join('categories as parents', 'parents.id', '=', 'categories.parent_id')
         //     ->select(['categories.*', 'parents.name as parent_name'])->get();
+
+        #todo how to use Roles (Gate)?
+        // if (!Gate::allows('categories.view')) {
+        //     abort(403);
+        // }
+        // if (Gate::denies('categories.view')) {
+        //     abort(403);
+        // }
+
+        #todo how to use the Policy?
+        $this->authorize('viewAny', Category::class);
 
         $categories = Category::leftJoin('categories as parents', 'parents.id', '=', 'categories.parent_id')
             ->select(['categories.*', 'parents.name as parent_name'])->paginate(10);
@@ -49,9 +66,6 @@ class CategoriesController extends Controller
             'categories' => $categories,
             'flash_message' => session('success')
         ]);
-        // return view('categories.index')->with([
-        //     'categories' => $entries
-        // ]);
     }
 
     // if the parameter in route is optional should add a default value
@@ -70,6 +84,13 @@ class CategoriesController extends Controller
 
     public function create()
     {
+        // if (Gate::denies('categories.create')) {
+        //     abort(403);
+        // }
+
+        #todo how to use the Policy?
+        $this->authorize('create', Category::class);
+
         $parents = Category::get();
         $category = new Category();
 
@@ -78,6 +99,13 @@ class CategoriesController extends Controller
 
     public function store(Request $request)
     {
+        // if (Gate::denies('categories.create')) {
+        //     abort(403);
+        // }
+
+        #todo how to use the Policy?
+        $this->authorize('store', Category::class);
+
         // DB::table('categories')->create([]);
 
         // $clean = $request->validate([
@@ -130,6 +158,9 @@ class CategoriesController extends Controller
         // $category = Category::where('id', '=', $id)->firstOrFail();
         // $category = Category::findOrFail($id);
 
+        #todo how to use the Policy?
+        $this->authorize('update', $category);
+
         $parents = Category::get();
         // $parents->pluck('name', 'id')->toArray();
         return view('categories.edit', ['category' => $category, 'parents' => $parents]);
@@ -144,6 +175,14 @@ class CategoriesController extends Controller
         //     'description' => ['nullable', 'string'],
         //     'art_path' => ['nullable', 'mimes:jpeg,jpg,png,gif,svg,webp']
         // ]);
+
+        // if (Gate::denies('categories.update')) {
+        //     abort(403);
+        // }
+
+        #todo how to use the Policy?
+        $this->authorize('update', $category);
+
         $clean = $request->validate($this->rules, $this->messages); // the message is a default value. not required
 
         // $category = Category::findOrFail($category->id);
@@ -162,8 +201,12 @@ class CategoriesController extends Controller
         // DB::table('categories')->where('id', '=', $id)->delete();
         // Category::where('id', '=', $id)->delete();
         // Category::destroy($id);
+
         $category = Category::findOrFail($id);
-        $category->delete();
+        #todo how to use the Policy?
+        $this->authorize('destroy', $category);
+
+        $category->destroy($category);
         // session()->flash('success', 'Category is Deleted');
         // Session::flash('success', 'Category is Deleted');
         return redirect()->route('categories.index')->with('success', 'Category is Deleted')->setStatusCode(200);
@@ -184,6 +227,10 @@ class CategoriesController extends Controller
     public function forceDelete($id)
     {
         $category = Category::withTrashed()->findOrFail($id);
+
+        #todo how to use the Policy?
+        $this->authorize('forceDelete', $category);
+
         $category->forceDelete();
         return to_route('categories.index')->with('success', 'Category is Deleted for ever!');
     }
